Included JWT security mechanism
Spring Security filter for handling JWTs in the authorization header has been implemented, along with the dependencies and components needed.

JWT validation and generation is done via the JwtUtil component using the JJWT library, for which an expiration time and secret key have to be set in the application.yml file.